+++ metta/agent/metta_agent.py
@@ -151,7 +151,9 @@ class MettaAgent(nn.Module):
         self.active_actions = list(zip(action_names, action_max_params, strict=False))
 
         # Precompute cumulative sums for faster conversion
-        self.cum_action_max_params = torch.cumsum(torch.tensor([0] + action_max_params, device=self.device), dim=0)
+        self.cum_action_max_params = torch.cumsum(
+            torch.tensor([0] + action_max_params, device=self.device, dtype=torch.long), dim=0
+        )
 
         full_action_names = []
         for action_name, max_param in self.active_actions:
